[feat] 删除便签API

+++ src/main/java/com/zouht/todolist/controller/note/NoteDeleteController.java
@@ -3,6 +3,7 @@ package com.zouht.todolist.controller.note;
 import com.zouht.todolist.service.note.NoteDeleteService;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import javax.annotation.Resource;
@@ -14,9 +15,15 @@ public class NoteDeleteController {
     NoteDeleteService noteDeleteService;
 
     @DeleteMapping("/note/delete")
-    public Map<String, Object> delete(@RequestBody Map<String, Object> map) {
-        Integer id = (Integer) map.get("id");
-        return noteDeleteService.delete(id);
-    }
+    public Map<String, Object> delete(@RequestParam Integer id) {
+        if (id == null) {
+            return Map.of("status", 1, "message", "id is null");
+        }
 
+        try {
+            return noteDeleteService.delete(id);
+        } catch (Exception e) {
+            return Map.of("status", 1, "message", e.getMessage());
+        }
+    }
 }
